% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_functions.R
\name{expand}
\alias{expand}
\title{expand}
\usage{
expand(weights)
}
\value{
A matrix of integers with one column per zone.
These serve as indices to create the final simulated data frame.
}
\description{
expand
}
\examples{
# not run
}

