% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_functions.R
\name{expand_zone}
\alias{expand_zone}
\title{expand_zone}
\usage{
expand_zone(zone)
}
\arguments{
\item{zone}{zone to iterate the expansion over}
}
\value{
A vector of integers per zone
}
\description{
expand_zone is a helper function that is run over each zone in the constraint
dataset and used in the \code{expand} function.
Typically you do not need to call expand_zone directly.
}
\examples{
# not run
}

