% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{extract}
\alias{extract}
\title{extract}
\usage{
extract(weights, inds, id)
}
\arguments{
\item{weights}{A weight table, typically produced by rakeR::weight()}

\item{inds}{The individual level data}

\item{id}{The unique id variable in the individual level data (inds),
usually the first column}
}
\value{
A data frame with zones and aggregated simulated values for each
variable
}
\description{
Deprecated. Use rk_extract() instead.
}
\examples{
\dontrun{
Deprecated. Use rk_extract()
}
}
