% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{extract_weights}
\alias{extract_weights}
\title{extract_weights}
\usage{
extract_weights(weights, inds, id)
}
\arguments{
\item{weights}{A weight table, typically produced using rakeR::rk_weight()}

\item{inds}{The individual level data}

\item{id}{The unique id variable in the individual level data (inds),
usually the first column}
}
\value{
A data frame with zones and aggregated simulated values for each
variable
}
\description{
Deprecated: use rakeR::rk_extract()
}
\examples{
\dontrun{
extract_weights() is deprecated, use rk_extract() instead
}
}
