% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{integerise}
\alias{integerise}
\alias{integerize}
\title{integerise}
\usage{
integerise(weights, inds, method = "trs", seed = 42)
}
\arguments{
\item{weights}{A matrix or data frame of fractional weights, typically
provided by \code{rakeR::weight()}}

\item{inds}{The individual--level data (i.e. one row per individual)}

\item{method}{The integerisation method specified as a character string.
Defaults to \code{"trs"}; currently other methods are not implemented.}

\item{seed}{The seed to use, defaults to 42.}
}
\value{
A data frame of integerised cases
}
\description{
Deprecated. Use rk_integerise()
}
\examples{
\dontrun{
cons <- data.frame(
  "zone"      = letters[1:3],
  "age_0_49"  = c(8, 2, 7),
  "age_gt_50" = c(4, 8, 4),
  "sex_f"     = c(6, 6, 8),
  "sex_m"     = c(6, 4, 3),
  stringsAsFactors = FALSE
)

inds <- data.frame(
  "id"     = LETTERS[1:5],
  "age"    = c("age_gt_50", "age_gt_50", "age_0_49", "age_gt_50", "age_0_49"),
  "sex"    = c("sex_m", "sex_m", "sex_m", "sex_f", "sex_f"),
  "income" = c(2868, 2474, 2231, 3152, 2473),
  stringsAsFactors = FALSE
)
vars <- c("age", "sex")

weights     <- rk_weight(cons = cons, inds = inds, vars = vars)
weights_int <- rk_integerise(weights, inds = inds)
}
}
