% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake.R
\name{rake}
\alias{rake}
\title{Rapid automatic extraction of keywords from documents}
\usage{
rake(x, split_words = smart_stop_words(), split_punct = basic_punct(),
  top_fraction = 1/3)
}
\arguments{
\item{x}{a character vector of texts to find keywords for}

\item{top_fraction}{the fraction of the most highly ranked phrases to return}

\item{method}{one of "degreeFreq", "degree", or "freq". degreeFreq is simply 
degree // frequency.}
}
\value{
returns a list with elements composed of one named integer vector for
each document. key phrases are names and values are ranked
}
\description{
This is the main functions for extracting and returning ranked keywords. This
algorithm works by tokenizing words and phrases based on stopwords. After 
tokenization the frequency of tokens and their degree \(i.e. their 
coocurrence with words in the remaining phrases\) are used to rank phrases. 
Stopwords can include both common english words and can be enhanced by 
including a vector of domain specific stop words.
}
\examples{
rake(test_text, 5, "degree")
rake(test_text, 15)
}
