% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_functions.R
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(weights, inds)
}
\arguments{
\item{weights}{A matrix of integerised weights provided by
\code{weight() \%>\% integerise()}.
One column per zone and one row per individual from \code{inds}}

\item{inds}{The individual--level data (i.e. one row per individual).
Ideally I would be able to pass this along the chain for you, but I don't
know how to do that yet so you must manually specify this again, sorry!}
}
\value{
A data frame with spatial microsimulated data, with one row per
(simulated) individual with an associated zone.
}
\description{
simulate
}
\examples{
# not run
}

