% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_in_place.R
\name{stem_in_place}
\alias{stem_in_place}
\title{stem but don't tokenize}
\usage{
stem_in_place(x, language = "english")
}
\arguments{
\item{x}{is a chr vector.}

\item{language}{should be one of “danish”, “dutch”, “english”, “finnish”,
“french”, “german”, “hungarian”, “italian”, “norwegian”, “porter”,
“portuguese”, “romanian”, “russian”, “spanish”, “swedish”, “turkish”}
}
\description{
a helper function to stem your chr vector and return stemmed vector this
reduces the total vocab significantly
}
