% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{weight}
\alias{weight}
\title{weight}
\usage{
weight(cons, inds, vars = NULL, iterations = 10)
}
\arguments{
\item{cons}{A data frame containing all the constraints. This
should be in the format of one row per zone, one column per constraint
category. The first column should be a zone code; all other columns must be
numeric counts.}

\item{inds}{A data frame containing individual-level (survey) data. This
should be in the format of one row per individual, one column per
constraint. The first column should be an individual ID.}

\item{vars}{A character vector of variables that constrain the simulation
(i.e. independent variables)}

\item{iterations}{The number of iterations the algorithm should complete.
Defaults to 10}
}
\value{
A data frame of fractional weights for each individual in each zone
with zone codes recorded in column names and individual id recorded in row
names.
}
\description{
Deprecated. Use rk_weight()
}
\examples{
# Deprecated. Use rk_weight()
}
