% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{check_constraint}
\alias{check_constraint}
\title{check_constraint}
\usage{
check_constraint(constraint_var, num_zones)
}
\arguments{
\item{constraint_var}{The constraint table to check, usually a data frame}

\item{num_zones}{The number of zones that should be present in the table}
}
\value{
If no errors are detected the function returns silently. Any errors
will stop the function or script to be investigated.
}
\description{
Deprecated: these checks are now done as part of the weight() and/or
extract()/integerise() steps automatically
}
\details{
Checks a constraint table for the following common errors:
\itemize{
  \item Ensures all zone codes are unique
  \item Ensures there are the expected number of zones
  \item Ensures all but the zone column are numeric (integer or double)
}
}
\examples{
Not run
## check_constraint() is deprecated. These checks are automatically
## carried out as part of the weight() and/or extract()/integerise()
## functions
}
