% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{check_ind}
\alias{check_ind}
\title{check_ind}
\usage{
check_ind(ind_var)
}
\arguments{
\item{ind_var}{the individual (survey) variable you want to check}
}
\value{
If no errors are detected the function returns silently. Any errors
will stop the function or script to be investigated.
}
\description{
Deprecated: these checks are now automatically carried out by
weight() and/or extract()/integerise()
}
\details{
Checks an individual (survey) variable for the following common errors:
\itemize{
  \item That each row sums to 1 (i.e. correctly converted to a dummy
  variable)
}
}
\examples{
## Not run
## checks are now automatically carried out by weight() and/or
## extract()/integerise()
}
