% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_functions.R
\name{extract_weights}
\alias{extract_weights}
\title{extract_weights}
\usage{
extract_weights(weights, inds, id)
}
\arguments{
\item{weights}{A weight table, typically produced by rakeR::weight()}

\item{inds}{The individual level data}

\item{id}{The unique id variable in the individual level data, usually the
first column}
}
\value{
A data frame with zones and aggregated simulated values for each
variable
}
\description{
Extract aggregate weights from individual weight table
}
\details{
Extract aggregate weights from individual weight table, typically produced
by rakeR::weight()

Extract cannot operate with numeric variables because it creates a new
variable for each unique factor of each variable
If you want numeric information, like income, use integerise() instead.
}
\examples{
## not run
}

