% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_functions.R
\name{extract_weights}
\alias{extract_weights}
\title{extract_weights}
\usage{
extract_weights(weights, inds, id)
}
\arguments{
\item{weights}{A weight table, typically produced by rakeR::weight()}

\item{inds}{The individual level data}

\item{id}{The unique id variable in the individual level data (inds),
usually the first column}
}
\value{
A data frame with zones and aggregated simulated values for each
variable
}
\description{
Deprecated: use rakeR::extract()
}
\examples{
## Not run
## extract_weights() is deprecated, use extract() instead
}
