% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_functions.R
\name{extract_weights}
\alias{extract_weights}
\title{extract_weights}
\usage{
extract_weights(weights, inds, id)
}
\arguments{
\item{weights}{A weight table, typically produced by rakeR::weight()}

\item{inds}{The individual level data}

\item{id}{The unique id variable in the individual level data (inds),
usually the first column}
}
\value{
A data frame with zones and aggregated simulated values for each
variable
}
\description{
Deprecated: use rakeR::extract()
}
\examples{
# SimpleWorld
cons <- data.frame(
"zone"   = letters[1:3],
"a0_49"  = c(8, 2, 7),
"a_gt50" = c(4, 8, 4),
"f"      = c(6, 6, 8),
"m"      = c(6, 4, 3),
stringsAsFactors = FALSE
)
inds <- data.frame(
"id"     = LETTERS[1:5],
"age"    = c("a_gt50", "a_gt50", "a0_49", "a_gt50", "a0_49"),
"sex"    = c("m", "m", "m", "f", "f"),
"income" = c("high", "low", "low", "high", "low"),
stringsAsFactors = FALSE
)

vars <- c("age", "sex")
weights <- weight(cons = cons, inds = inds, vars = vars)

ext_weights <- extract_weights(weights, inds = inds, id = "id")
print(ext_weights)
}
