% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{integerise}
\alias{integerise}
\alias{integerize}
\title{integerise}
\usage{
integerise(weights, inds, method = "trs", seed = 42)
}
\arguments{
\item{weights}{A matrix or data frame of fractional weights, typically
provided by \code{rakeR::weight()}}

\item{inds}{The individual--level data (i.e. one row per individual)}

\item{method}{The integerisation method specified as a character string.
Defaults to \code{"trs"}; currently other methods are not implemented.}

\item{seed}{The seed to use, defaults to 42.}
}
\value{
A data frame of integerised cases
}
\description{
Deprecated. Use rk_integerise()
}
\examples{
\dontrun{
Deprecated. Use rk_integerise()
}
}
