% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_functions.R
\name{integerise}
\alias{integerise}
\title{Integerise}
\usage{
integerise(weights, method = "trs")
}
\arguments{
\item{weights}{A weights matrix, typically provided by \code{weight()}}

\item{method}{The integerisation method specified as a character.
Defaults to \code{"trs"}.}
}
\value{
A matrix of integerised weights to be used by \code{expand()}
}
\description{
Generate integer cases from numeric weights matrix.
}
\details{
The default integerisation method uses the 'truncate, replicate, sample'
method developed by Robin Lovelace and Dimitris Ballas
\url{http://www.sciencedirect.com/science/article/pii/S0198971513000240}

Other methods (for example proportional probabilities) may be implemented
at a later date.
}
\examples{
# not run
}

