% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake.R
\name{rake}
\alias{rake}
\title{Rapid automatic extraction of keywords from documents}
\usage{
rake(x, n = 10, method = "degreeFreq")
}
\arguments{
\item{x}{a character vector of texts to find keywords for}

\item{n}{the number of ranked keyphrases to return}

\item{metho}{one of "degreeFreq", "degree", or "freq". degreeFreq is simply
degree // frequency.}
}
\description{
This is the main functions for extracting and returning ranked keywords. This
algorithm works by tokenizing words and phrases based on stopwords. After 
tokenization the frequency of tokens and their degree \(i.e. their 
coocurrence with words in the remaining phrases\) are used to rank phrases. 
Stopwords can include both common english words and can be enhanced by 
including a vector of domain specific stop words.
}
\examples{
rake(test_text, 5, "degree")
rake(test_text, 15)  
}

