% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_functions.R
\name{rake}
\alias{rake}
\title{rake}
\usage{
rake(cons, ind, vars)
}
\arguments{
\item{cons}{A data frame or matrix containing all the constraints.}

\item{ind}{A data frame containing individual--level (survey) data.}

\item{vars}{A vector or list or variables that constrain the simulation (i.e. independent variables)}
}
\value{
A data frame of fractional weights for each individual in each zone.
}
\description{
Produces fractional weights using the iterative proportional fitting ('raking') algorithm.
It takes two data frames or matrices as arguments, one for the constraints and one for the individual--level survey data.
The constraints should be one row per zone, one column per constraint category.
The first column should contain the zone codes.
The survey data should be one row per individual, one column per constraint.
The first column should contain the individual unique identifier.
Both data frames should only contain an ID column (zone ID, individual ID) and constraints or constraint category.
No other columns should be present (the user can merge these back in later).
This is typically the format you will obtain and/or tidy data into, so the function does all the necessary re-structuring.
The fractional weights can be integerised to create a simulated data frame.
}
\examples{
# SimpleWorld
cons <- data.frame(
"zone"  = 1:3,
"a0_49" = c(8, 2, 7),
"a_50+" = c(4, 8, 4),
"f"    = c(6, 6, 8)
"m"    = c(6, 4, 3),
)
ind <- data.frame(
"id"     = 1:5,
"age"    = c("a_50+", "a_50+", "a0_49", "a_50+", "a0_49"),
"sex"    = c("m", "m", "m", "f", "f"),
"income" = c(2868, 2474, 2231, 3152, 2473),
stringsAsFactors = FALSE
)
# Set variables to constrain over
vars <- c("age", "sex")
rake(cons = cons, ind = ind, vars = vars)
}

