% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_functions.R
\name{rake}
\alias{rake}
\title{rake}
\usage{
rake(cons, inds, vars, iterations = 10, method = "trs", ...)
}
\arguments{
\item{cons}{A data frame of constraint variables}

\item{inds}{A data frame of individual--level (survey) data}

\item{vars}{A character string of variables to iterate over}

\item{iterations}{The number of iterations to perform. Defaults to 10.}

\item{method}{Integerisation method to apply. Defaults to \code{trs}.}

\item{...}{Additional arguments to pass to additional methods}
}
\value{
A data frame of simulated individuals in zones.
}
\description{
A convenience function wrapping \code{weight()}, \code{integerise} and
\code{simulate}
}
\examples{
# not run
}

