% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_functions.R
\name{rake}
\alias{rake}
\title{rake}
\usage{
rake(cons, ind, vars)
}
\arguments{
\item{cons}{A data frame or matrix containing all the constraints. This
should be in the format of one row per zone, one column per constraint
category.}

\item{ind}{A data frame containing individual--level (survey) data. This
should be in the format of one row per individual, one column per
constraint.}

\item{vars}{A vector or list or variables that constrain the simulation (i.e.
independent variables)}
}
\value{
A data frame of fractional weights for each individual in each zone.
}
\description{
Produces fractional weights using the iterative proportional fitting
('raking') algorithm.
}
\details{
The first column of each data frame should be an ID. The first column of
\code{cons} should contain the zone codes. The first column of \code{ind}
should contain the individual unique identifier.

Both data frames should only contain:
\itemize{
  \item an ID column (zone ID \code{cons} or individual ID \code{ind}).
  \item constraints \code{ind} or constraint category \code{cons}.
  \item \code{ind} can optionally contain additional dependent variables
  that do not influence the weighting process.
}

No other columns should be present (the user can merge these back in later).

It is essential that the levels in each \code{ind} constraint (i.e. column)
match exactly with the column names in \code{cons}. In the example below see
how the column names in cons (\code{'a0_49', 'f', ...}) match exactly the
levels in \code{ind} variables.

The columns in \code{cons} must be in alphabetical order because these are
created alphabetically when they are 'spread' in the individual--level data.
}
\examples{
# SimpleWorld
cons <- data.frame(
"zone"  = 1:3,
"a0_49" = c(8, 2, 7),
"a_50+" = c(4, 8, 4),
"f"    = c(6, 6, 8),
"m"    = c(6, 4, 3)
)
ind <- data.frame(
"id"     = 1:5,
"age"    = c("a_50+", "a_50+", "a0_49", "a_50+", "a0_49"),
"sex"    = c("m", "m", "m", "f", "f"),
"income" = c(2868, 2474, 2231, 3152, 2473),
stringsAsFactors = FALSE
)
# Set variables to constrain over
vars <- c("age", "sex")
weights <- rake(cons = cons, ind = ind, vars = vars)
print(weights)
}

