% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(...)
}
\arguments{
\item{...}{arguments previously passed to simulate()}
}
\value{
Returns an error if used. Just use integerise()
}
\description{
Deprecated: integerise() \%>\% simulate() has been replaced by simply
integerise() to be consistent with extract().
}
