% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_functions.R
\name{trs}
\alias{trs}
\title{trs}
\usage{
trs(weights)
}
\arguments{
\item{weights}{a weights matrix, typically provided by \code{rake()}}
}
\value{
A matrix of integer weights
}
\description{
Truncate, replicate, sample: a helper function to generate integer cases
from numeric weights.
}
\details{
Truncate, replicate, sample is a method of integerisation developed by Robin
Lovelace and Dimitris Ballas
\url{http://www.sciencedirect.com/science/article/pii/S0198971513000240}
}
\examples{
# not run
}

