% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{weight}
\alias{weight}
\title{weight}
\usage{
weight(cons, inds, vars = NULL, iterations = 10)
}
\arguments{
\item{cons}{A data frame containing all the constraints. This
should be in the format of one row per zone, one column per constraint
category. The first column should be a zone code; all other columns must be
numeric counts.}

\item{inds}{A data frame containing individual-level (survey) data. This
should be in the format of one row per individual, one column per
constraint. The first column should be an individual ID.}

\item{vars}{A character vector of variables that constrain the simulation
(i.e. independent variables)}

\item{iterations}{The number of iterations the algorithm should complete.
Defaults to 10}
}
\value{
A data frame of fractional weights for each individual in each zone
with zone codes recorded in column names and individual id recorded in row
names.
}
\description{
Deprecated. Use rk_weight()
}
\examples{
\dontrun{
# SimpleWorld
cons <- data.frame(
  "zone"      = letters[1:3],
  "age_0_49"  = c(8, 2, 7),
  "age_gt_50" = c(4, 8, 4),
  "sex_f"     = c(6, 6, 8),
  "sex_m"     = c(6, 4, 3),
  stringsAsFactors = FALSE
)
inds <- data.frame(
  "id"     = LETTERS[1:5],
  "age"    = c("age_gt_50", "age_gt_50", "age_0_49", "age_gt_50", "age_0_49"),
  "sex"    = c("sex_m", "sex_m", "sex_m", "sex_f", "sex_f"),
  "income" = c(2868, 2474, 2231, 3152, 2473),
  stringsAsFactors = FALSE
)
# Set variables to constrain over
vars <- c("age", "sex")
weights <- weight(cons = cons, inds = inds, vars = vars)
print(weights)
}
}
